refactor: modify entity
