+++ src/main/java/com/nhnacademy/inkbridge/backend/entity/BackendApplication.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.inkbridge.backend;
+package com.nhnacademy.inkbridge.backend.entity;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
