--- src/main/java/com/nhnacademy/inkbridge/backend/entity/CouponStatus.java
@@ -1,31 +1,0 @@
-package com.nhnacademy.inkbridge.backend.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * class: CouponStatus.
- *
- * @author nhn
- * @version 2024/02/08
- */
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Entity
-@Table(name = "coupon_status")
-public class CouponStatus {
-    @Id
-    @Column(name = "status_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long statusId;
-
-    @Column(name = "status_name")
-    private String statusName;
-}
