+++ src/main/java/com/nhnacademy/inkbridge/backend/entity/ReviewFile.java
@@ -1,0 +1,21 @@
+package com.nhnacademy.inkbridge.backend.entity;
+
+import javax.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "review_file")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ReviewFile {
+    @Id
+    @OneToOne
+    @JoinColumn(name = "file_id")
+    private File file;
+
+    @ManyToOne
+    @JoinColumn(name = "review_id")
+    private Book book;
+}
