+++ src/main/java/com/nhnacademy/inkbridge/backend/entity/User.java
@@ -4,16 +4,7 @@ import com.nhnacademy.inkbridge.backend.entity.enums.Auth;
 import com.nhnacademy.inkbridge.backend.entity.enums.Status;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
+import javax.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -66,8 +57,11 @@ public class User {
     @Enumerated(EnumType.STRING)
     private Status status;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "grade_id")
     private UserGrade userGrade;
+
+    @Column(name = "withdraw_at")
+    private LocalDateTime withdrawAt;
 
 }
