+++ src/main/java/com/nhnacademy/inkbridge/backend/entity/Wrapping.java
@@ -31,4 +31,7 @@ public class Wrapping {
 
     @Column(name = "price")
     private Long price;
+
+    @Column(name = "is_active")
+    private Boolean isActive;
 }
